Move RoadInfoIterator to its own header
